% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWCD-package.R
\docType{package}
\name{calcWCD-package}
\alias{calcWCD}
\alias{calcWCD-package}
\title{calcWCD: Calculates WCD}
\description{
Calculate weighted co-publication distance, a measure of semantic similarity, between a literature topic and a gene/protein.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ed-lau/calcWCD}
  \item Report bugs at \url{https://github.com/ed-lau/calcWCD/issues}
}

}
\author{
\strong{Maintainer}: Edward Lau \email{lau1@stanford.edu}

}
\keyword{internal}
