% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmid.R
\name{pmid}
\alias{pmid}
\title{pmid}
\usage{
pmid(q)
}
\arguments{
\item{q}{character() A character string specifying search topic}
}
\description{
Retrieves a list of PMIDs from the NCBI EUtils API based on a search term
}
\examples{
pmid("fibrosis NOT cystic")
}
\keyword{PubMed}
\keyword{query}
