% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{wcd}
\alias{wcd}
\title{wcd}
\usage{
wcd(pmids, annotation)
}
\arguments{
\item{pmids}{A one-column dataframe containing the total list of PMID accessions in the topic under consideration. A test file is provided in test_pmids.Rds for format requirement.}

\item{annotation}{A tibble of GeneID x PMID relationships, cite counts, and pub year. A  test file containing 100,000 relationships is provided in test_annotation.Rds for format requirement.}
}
\description{
Calculate weighted co-publication distance given a list of PubMed IDs (PMIDs) and an annotation file containing GeneID and PMID links.
}
\examples{
wcd(pmids = test_pmids, annotation = test_annotation)
}
\keyword{semantic}
\keyword{similarity}
